Avoid segfault in backends destructor, when terminate related process.

+++ src/backends/R/rsession.cpp
@@ -39,7 +39,8 @@ RSession::RSession(Cantor::Backend* backend) : Session(backend), m_process(nullp
 RSession::~RSession()
 {
     qDebug();
-    m_process->terminate();
+    if (m_process)
+        m_process->terminate();
 }
 
 void RSession::login()
